packages/create-sidclaw-app/templates/plain-python/requirements.txt
sidclaw>=0.1.0
python-dotenv>=1.0.0
